Changing Abstraction and inheritance structure to incorporate Factory or Bridge pattern
